+++ src/main/templates/net/imagej/ops/Ops.list
@@ -61,7 +61,6 @@ namespaces = ```
 		[name: "imgLabeling",                 iface: "ImgLabeling"],
 		[name: "imgPlus",                     iface: "ImgPlus"],
 		[name: "integerType",                 iface: "IntegerType"],
-		[name: "integralImg",                 iface: "IntegralImg"],
 	[name: "kernel",                      iface: "Kernel"],
 		[name: "kernelGauss",                 iface: "KernelGauss"],
 		[name: "kernelLog",                   iface: "KernelLog"],
@@ -163,6 +162,7 @@ namespaces = ```
 		[name: "equation",                    iface: "Equation"],
 		[name: "fill",                        iface: "Fill"],
 		[name: "histogram",                   iface: "Histogram"],
+		[name: "integral",                    iface: "Integral"],
 		[name: "invert",                      iface: "Invert"],
 		[name: "normalize",                   iface: "Normalize",           aliases: ["norm"]],
 		[name: "project",                     iface: "Project"],
